+++ router/core/router.go
@@ -6,8 +6,6 @@ import (
 	"crypto/x509"
 	"errors"
 	"fmt"
-	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/propagation"
 	"net"
 	"net/http"
 	"net/url"
@@ -16,6 +14,9 @@ import (
 	"strings"
 	"sync"
 	"time"
+
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/propagation"
 
 	"github.com/wundergraph/graphql-go-tools/v2/pkg/netpoll"
 
@@ -133,6 +134,7 @@ type (
 	RouterConfigPollerConfig struct {
 		config.ExecutionConfig
 		PollInterval time.Duration
+		PollJitter   time.Duration
 		GraphSignKey string
 	}
 
@@ -179,6 +181,7 @@ type (
 		healthCheckPath                 string
 		readinessCheckPath              string
 		livenessCheckPath               string
+		playgroundConfig                config.PlaygroundConfig
 		cacheControlPolicy              config.CacheControlPolicy
 		routerConfigPollerConfig        *RouterConfigPollerConfig
 		cdnConfig                       config.CDNConfiguration
@@ -260,8 +263,18 @@ func NewRouter(opts ...Option) (*Router, error) {
 		r.graphqlWebURL = r.graphqlPath
 	}
 
-	if r.playgroundPath == "" {
-		r.playgroundPath = "/"
+	// this is set via the deprecated method
+	if !r.playground {
+		r.playgroundConfig.Enabled = r.playground
+		r.logger.Warn("The playground_enabled option is deprecated. Use the playground.enabled option in the config instead.")
+	}
+	if r.playgroundPath != "" && r.playgroundPath != "/" {
+		r.playgroundConfig.Path = r.playgroundPath
+		r.logger.Warn("The playground_path option is deprecated. Use the playground.path option in the config instead.")
+	}
+
+	if r.playgroundConfig.Path == "" {
+		r.playgroundConfig.Path = "/"
 	}
 
 	if r.instanceID == "" {
@@ -600,7 +613,7 @@ func (r *Router) newServer(ctx context.Context, cfg *nodev1.RouterConfig) error 
 func (r *Router) listenAndServe(cfg *nodev1.RouterConfig) error {
 	r.logger.Info("Server listening and serving",
 		zap.String("listen_addr", r.listenAddr),
-		zap.Bool("playground", r.playground),
+		zap.Bool("playground", r.playgroundConfig.Enabled),
 		zap.Bool("introspection", r.introspection),
 		zap.String("config_version", cfg.GetVersion()),
 	)
@@ -622,7 +635,6 @@ func (r *Router) listenAndServe(cfg *nodev1.RouterConfig) error {
 }
 
 func (r *Router) initModules(ctx context.Context) error {
-
 	moduleList := make([]ModuleInfo, 0, len(modules)+len(r.customModules))
 
 	for _, module := range modules {
@@ -857,15 +869,16 @@ func (r *Router) bootstrap(ctx context.Context) error {
 		debug.ReportMemoryUsage(ctx, r.logger)
 	}
 
-	if r.playground {
-		playgroundUrl, err := url.JoinPath(r.baseURL, r.playgroundPath)
+	if r.playgroundConfig.Enabled {
+		playgroundUrl, err := url.JoinPath(r.baseURL, r.playgroundConfig.Path)
 		if err != nil {
 			return fmt.Errorf("failed to join playground url: %w", err)
 		}
 		r.logger.Info("Serving GraphQL playground", zap.String("url", playgroundUrl))
 		r.playgroundHandler = graphiql.NewPlayground(&graphiql.PlaygroundOptions{
-			Html:       graphiql.PlaygroundHTML(),
-			GraphqlURL: r.graphqlWebURL,
+			Html:             graphiql.PlaygroundHTML(),
+			GraphqlURL:       r.graphqlWebURL,
+			ConcurrencyLimit: int64(r.playgroundConfig.ConcurrencyLimit),
 		})
 	}
 
@@ -978,7 +991,6 @@ func (r *Router) buildClients() error {
 			StorageConfig: &provider,
 			Prefix:        r.automaticPersistedQueriesConfig.Storage.ObjectPrefix,
 		})
-
 		if err != nil {
 			return err
 		}
@@ -1164,7 +1176,7 @@ func (r *Router) Start(ctx context.Context) error {
 		return err
 	}
 
-	if r.playground {
+	if r.playgroundConfig.Enabled {
 		graphqlEndpointURL, err := url.JoinPath(r.baseURL, r.graphqlPath)
 		if err != nil {
 			return fmt.Errorf("failed to join graphql endpoint url: %w", err)
@@ -1220,7 +1232,6 @@ func (r *Router) Start(ctx context.Context) error {
 // Shutdown gracefully shuts down the router. It blocks until the server is shutdown.
 // If the router is already shutdown, the method returns immediately without error.
 func (r *Router) Shutdown(ctx context.Context) (err error) {
-
 	if !r.shutdown.CompareAndSwap(false, true) {
 		return nil
 	}
@@ -1415,6 +1426,13 @@ func WithGraphQLWebURL(p string) Option {
 func WithPlaygroundPath(p string) Option {
 	return func(r *Router) {
 		r.playgroundPath = p
+	}
+}
+
+// WithPlaygroundPath sets the path where the GraphQL Playground is served.
+func WithPlaygroundConfig(c config.PlaygroundConfig) Option {
+	return func(r *Router) {
+		r.playgroundConfig = c
 	}
 }
 
@@ -1909,7 +1927,6 @@ func buildAttributesMap(attributes []config.CustomAttribute) map[string]string {
 
 // buildHeaderAttributesMapper returns a function that maps custom attributes to the request headers.
 func buildHeaderAttributesMapper(attributes []config.CustomAttribute) func(req *http.Request) []attribute.KeyValue {
-
 	if len(attributes) == 0 {
 		return nil
 	}